srm: Move log4j to Tomcat common/lib directory
Tomcat, JGlobus or PureTLS use Apache commons logging. Commons logging
auto discovers available logging systems, such as log4j. However log4j
was deployed as part of the SRM webapp, not as part of tomcat. Hence
several log targets would map to Jdk14Logger rather than to the log4j
logger.

This patch makes sure that the log4j jar is loaded as part of tomcat
rather than as part of the SRM webapp. This exposes additional logging
targets.

Target: trunk
Target: 1.9.5
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1218/

+++ modules/dCache/diskCacheV111/srm/dcache/Storage.java
@@ -3184,7 +3184,7 @@ public class Storage
     @Override
     public List<FileMetaData>
         listDirectory(SRMUser user, String directory, final boolean verbose,
-                      int offset, int count)
+                      long offset, long count)
         throws SRMException
     {
         final Set<FileAttribute> required =
